fix: rename ClockOps to Clock
